+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 ### Features
 
 - richat: rm `solana-sdk` ([#174](https://github.com/lamports-dev/richat/pull/174))
+- richat: use channel per source ([#175](https://github.com/lamports-dev/richat/pull/175))
 
 ### Breaking